Move and rename RoomCallNavStatus
